requirements.txt
# Core dependencies
flask>=2.0.0
flask-socketio>=5.1.1
python-socketio>=5.4.0
aiohttp>=3.8.0
aiojobs>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.0
pyshark>=0.6.0

# Security tools
python-nmap>=0.7.1
yara-python>=4.3.0
cryptography>=41.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# System monitoring
psutil>=5.9.0
wmi>=1.5.1
pywin32>=306; platform_system == "Windows"
watchdog>=2.1.7

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.7.3

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
dash>=2.13.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Template Engine
aiohttp-jinja2>=1.5.0
jinja2>=3.1.0

# Additional dependencies
python-dotenv>=1.0.0
prometheus-client>=0.17.0
